KEYCLOAK-2834 Migrate org.keycloak.testsuite.admin tests to Arquillian
