+++ testsuite/integration-arquillian/tests/base/src/test/resources/client-descriptions/client-oidc.json
@@ -1,0 +1,6 @@
+{
+  "client_name": "Name",
+  "redirect_uris": [
+    "http://localhost"
+  ]
+}
